refactor(users): update user view models and validation
* refactor create and edit user view models to include validation attributes
* remove obsolete view models
* update localization for user creation and editing
* adjust namespaces for consistency
