+++ src/Beagl.WebApp/Pages/Users/Edit.cshtml
@@ -40,14 +40,12 @@
                                     @Localizer["Username"] <span class="text-danger">*</span>
                                 </label>
                                 <input asp-for="Input.UserName" class="form-control" />
-                                <span asp-validation-for="Input.UserName" class="text-danger"></span>
                             </div>
                             <div class="mb-3">
                                 <label class="form-label">
-                                    @Localizer["Email"]
+                                    @Localizer["Email"] <span class="text-danger">*</span>
                                 </label>
                                 <input asp-for="Input.Email" class="form-control" />
-                                <span asp-validation-for="Input.Email" class="text-danger"></span>
                             </div>
                         </div>
                         <div class="col-md-6">
@@ -56,7 +54,6 @@
                                     @Localizer["Phone Number"] <span class="text-danger">*</span>
                                 </label>
                                 <input asp-for="Input.PhoneNumber" class="form-control" />
-                                <span asp-validation-for="Input.PhoneNumber" class="text-danger"></span>
                             </div>
                             <div class="mb-3">
                                 <label class="form-label">
@@ -91,7 +88,6 @@
                                         }
                                     }
                                 </select>
-                                <span asp-validation-for="Input.Roles" class="text-danger"></span>
                             </div>
                         </div>
                     </div>
