+++ src/Beagl.WebApp/Resources/Pages/Users/Create.fr.resx
@@ -153,4 +153,25 @@
   <data name="Select Role" xml:space="preserve">
     <value>Sélectionner un rôle</value>
   </data>
+  <data name="EmailFormat_Invalid" xml:space="preserve">
+    <value>Le format du courriel est invalide.</value>
+  </data>
+  <data name="PhoneFormat_Invalid" xml:space="preserve">
+    <value>Le format du numéro de téléphone est invalide.</value>
+  </data>
+  <data name="UserName_Required" xml:space="preserve">
+    <value>Le nom d'utilisateur est requis.</value>
+  </data>
+  <data name="Password_Required" xml:space="preserve">
+    <value>Le mot de passe est requis.</value>
+  </data>
+  <data name="ConfirmPassword_Required" xml:space="preserve">
+    <value>La confirmation du mot de passe est requise.</value>
+  </data>
+  <data name="Phone_Required" xml:space="preserve">
+    <value>Le numéro de téléphone est requis.</value>
+  </data>
+  <data name="Email_Required" xml:space="preserve">
+    <value>Le courriel est requis.</value>
+  </data>
 </root>
